Refactor to move sql slash into it's own file
